backend/requirements.txt
# YAAS
openai
python-dotenv
tiktoken
nltk
konlpy
git+https://github.com/allinux/PyKoSpacing.git
# Web Framework, DB
fastapi
tzdata
psycopg2-binary
asyncpg
sqlalchemy
alembic
pydantic
uvicorn
bcrypt